Summary of all changes Bugs fixed Bug 1 — EtherDream rate truncation (EtherDreamController.hpp, EtherDreamController.cpp): pendingRateChanges, nextPendingRateChange(), and sendPointRate() all promoted from uint16_t to uint32_t. Point rates above 65535 pps now transmit correctly.
Bug 2 — logError for normal EtherDream state transitions (EtherDreamController.cpp): sendPrepare(), sendBegin(), and sendPointRate() now use logInfoVerbose instead of logError.

Bug 3 — connect() timeout ordering (EtherDreamController.cpp): setDefaultTimeout(200ms) and setConnectTimeout(1s) moved to before the connect() call so the first connection attempt uses the intended timeout.

Bug 4 — missing resetStartupBlank() on reconnect (HeliosController.cpp, LaserCubeNetController.cpp): Helios now calls it when transitioning from disconnected → connected; LaserCubeNet calls it at the start of run() matching the other controllers.

Bug 5 — LaserCubeUSB min==max forces blank padding (LaserCubeUsbController.cpp): minimumPointsRequired changed to 1, allowing the frame queue to emit natural frame boundaries without forcing exact-count padding.

Bug 6 — infinite USB timeouts (LaserCubeUsbController.cpp): All libusb_bulk_transfer calls in control helper functions now use 1000ms timeout. Data endpoint uses 100ms per attempt (3 retries = 300ms max), making the previous dead-code retry loop actually functional.

Bug 7 — assert + redundant guard (LaserController.cpp): Removed the assert that would crash in debug builds; kept the clean if (frame.points.size() == 0) return false; guard.

Performance improvements
Perf 1 — bulk copy in fillFromFrameQueue (LaserController.cpp): Point-by-point push_back loop replaced with a single insert() over a contiguous range. The compiler can now vectorise this in the realtime hot path.

Perf 2 — messageTimes fixed array (LaserCubeNetController.hpp, LaserCubeNetController.cpp): std::map<uint8_t, time_point> replaced with std::array<time_point, 256>. O(1) access/insert/erase, zero heap allocation, pendingAckCount tracks occupancy.

Perf 3 — latency stats cache (LaserControllerStreaming.hpp, LaserControllerStreaming.cpp): getLatencyStats() now skips the vector allocation + sort if no new samples have arrived since the last call (mutation counter check).

Perf 4 — lock released before frame queue metrics (LaserController.cpp): updateFrameQueueMetricsUnsafe() (which iterates frameQueue) is now called outside the pendingFramesMutex lock, reducing contention with the producer thread's sendFrame().

Perf 5 — adaptive sleep in LaserCubeNet (LaserCubeNetController.cpp): The fixed 1ms sleep replaced with a buffer-aware sleep: waits approximately as long as the current buffer excess takes to drain, clamped to 1–10ms. Reduces CPU usage significantly when the buffer is healthy.

Naming fixes
getArmed() → isArmed() in base class (LaserControllerStreaming.hpp)
frameModeEnabled() → isFrameModeEnabled() (LaserController.hpp)
frameFillCallback() → fillFromFrameQueue() (LaserController.hpp)
pps/newPps → currentPointRate/pendingPointRate in LaserCubeNet (LaserCubeNetController.hpp)
currentPps/targetPps → currentPointRate/pendingPointRate in LaserCubeUSB (LaserCubeUsbController.hpp)
UsbControllerHandle → LaserCubeUsbHandle (LaserCubeUsbController.hpp)
src/EtherDream/ directory: lowercase rename is a no-op on macOS (case-insensitive filesystem) — skipped
Other improvements
Frame shared_ptr → unique_ptr (LaserController.hpp, LaserController.cpp): pendingFrames and frameQueue now use unique_ptr<Frame>, removing atomic ref-count overhead from the realtime path. fillFromFrameQueue uses a raw Frame* pointer so pop_front() doesn't require extending frame lifetime.
Stale ack cleanup threshold (LaserCubeNetController.cpp): Reduced from 10 seconds to 1 second, matching the warning threshold and avoiding stale timestamps corrupting RTT measurements on message-number wrap.
LaserCubeUSB magic byte comments (LaserCubeUsbController.cpp): Init sequence documented as runner_mode_enable/disable, runner_mode_run/stop, runner_mode_load (confirmed from LaserdockDevice.cpp in oldcode).

+++ include/libera/etherdream/EtherDreamController.hpp
@@ -69,7 +69,7 @@ protected:
 
 
 private:
-    std::optional<std::uint16_t> nextPendingRateChange();
+    std::optional<std::uint32_t> nextPendingRateChange();
 
     /// Wait for the response frame to a specific command.
     expected<Ack>
@@ -81,7 +81,7 @@ private:
 
     /// Issue the point-rate command ('q') and return the associated ACK.
     expected<Ack>
-    sendPointRate(std::uint16_t rate);
+    sendPointRate(std::uint32_t rate);
 
     std::size_t calculateMinimumPoints();
     void sleepUntilNextPoints();
@@ -123,7 +123,7 @@ private:
     std::optional<std::error_code> lastError;
 
     std::mutex pendingRatesMutex;
-    std::deque<std::uint16_t> pendingRateChanges;
+    std::deque<std::uint32_t> pendingRateChanges;
     size_t pendingRateChangeCount = 0; 
 
     mutable std::atomic<int> lastEstimatedBufferFullness{0};
